refactor(agent): fold licensing error mapping into exceptions/llm
